dj - separate auditing handler, wiremock server from FrontiersMain
