+++ pom.xml
@@ -335,6 +335,7 @@
             <param>${app.package}.config.SecurityConfig</param>
             <param>${app.package}.config.SpaCsrfTokenRequestHandler</param>
             <param>${app.package}.config.CsrfCookieFilter</param>
+            <param>${app.package}.config.JpaAuditingConfig</param>
             <param>edu.ucsb.cs156.frontiers.services.wiremock.WiremockService</param>
             <param>edu.ucsb.cs156.frontiers.services.wiremock.WiremockServiceDummy</param>
             <param>edu.ucsb.cs156.frontiers.services.wiremock.WiremockServiceImpl</param>
